refactor: move download tab into programme view
